Update the min and max dates for the calender picker
Customers were previously able to select an arrival date that was before the start of the rental. And the same for departing. This fix provides a solution to that by blocking off the arrival and departure dates and only allowing dates after the start date or before the end date to be updates by the customer.